feat: Refactor item create and edit flows
- Extract ItemCreatePage from MainContent for cleaner separation
- Update ItemEditDialog to handle dynamic attribute fields and validation
- Modify ItemRepository to support inserting items with attributes
- Update ItemViewModel to handle item creation and editing with attributes
- Improve UI and UX for item creation and editing
