+++ app/src/main/java/com/example/itemanagerv2/MainActivity.kt
@@ -25,8 +25,7 @@ import kotlinx.coroutines.launch
 import com.example.itemanagerv2.data.local.entity.CategoryAttribute
 import com.example.itemanagerv2.data.local.model.ItemCardDetail
 import com.example.itemanagerv2.data.local.model.ItemCategoryArg
-import com.example.itemanagerv2.ui.component.CategoryListPage
-import com.example.itemanagerv2.ui.component.MainPage
+import com.example.itemanagerv2.ui.component.*
 import com.example.itemanagerv2.ui.theme.BaseTheme
 import com.example.itemanagerv2.viewmodel.ItemViewModel
 import dagger.hilt.android.AndroidEntryPoint
@@ -34,21 +33,19 @@ import android.graphics.BitmapFactory
 import android.content.pm.PackageManager
 import androidx.activity.OnBackPressedCallback
 import androidx.core.content.ContextCompat
-import com.example.itemanagerv2.ui.component.ItemEditDialog
-import com.example.itemanagerv2.ui.component.ItemViewDialog
 
 @AndroidEntryPoint
 class MainActivity : ComponentActivity() {
     private val itemViewModel: ItemViewModel by viewModels()
     private var currentEditingItemId: Int = 0
     private var pendingImagePick: Boolean = false
-    private var showAddDialog = mutableStateOf(false)
     private var selectedTab = mutableStateOf(0)
     private var onImageAddedCallback: ((ItemCardDetail) -> Unit)? = null
     private var selectedItem = mutableStateOf<ItemCardDetail?>(null)
     private var showEditDialog = mutableStateOf(false)
     private var isEditFromView = mutableStateOf(false)
     private var forceRecompose = mutableStateOf(0)
+    private var showCreatePage = mutableStateOf(false)
 
     private val imagePickerLauncher = registerForActivityResult(
         ActivityResultContracts.GetContent()
@@ -100,7 +97,6 @@ class MainActivity : ComponentActivity() {
             try {
                 itemViewModel.refreshItems()
                 
-                // If we're on the categories tab, also refresh the category attributes
                 if (selectedTab.value == 1) {
                     val currentCategory = itemViewModel.categories.value.firstOrNull()
                     currentCategory?.let { category ->
@@ -108,7 +104,6 @@ class MainActivity : ComponentActivity() {
                     }
                 }
                 
-                // Force a recomposition
                 forceRecompose.value = forceRecompose.value + 1
                 Log.d("MainActivity", "Data refresh complete")
             } catch (e: Exception) {
@@ -123,8 +118,8 @@ class MainActivity : ComponentActivity() {
         onBackPressedDispatcher.addCallback(this, object : OnBackPressedCallback(true) {
             override fun handleOnBackPressed() {
                 when {
-                    showAddDialog.value -> {
-                        showAddDialog.value = false
+                    showCreatePage.value -> {
+                        showCreatePage.value = false
                     }
                     showEditDialog.value -> {
                         showEditDialog.value = false
@@ -140,7 +135,6 @@ class MainActivity : ComponentActivity() {
                         selectedTab.value = 0
                     }
                     selectedTab.value == 0 -> {
-                        // Only finish() when on main screen with no dialogs or selected items
                         finish()
                     }
                 }
@@ -149,7 +143,6 @@ class MainActivity : ComponentActivity() {
 
         setContent {
             BaseTheme {
-                // Use forceRecompose to trigger recomposition
                 forceRecompose.value
                 
                 MainContent(
@@ -164,7 +157,7 @@ class MainActivity : ComponentActivity() {
                             itemViewModel.refreshItems()
                         }
                     },
-                    showAddDialog = showAddDialog,
+                    showCreatePage = showCreatePage,
                     selectedTab = selectedTab,
                     selectedItem = selectedItem,
                     showEditDialog = showEditDialog,
@@ -179,7 +172,7 @@ class MainActivity : ComponentActivity() {
 fun MainContent(
     itemViewModel: ItemViewModel,
     onPickImage: (Int, (ItemCardDetail) -> Unit) -> Unit,
-    showAddDialog: MutableState<Boolean>,
+    showCreatePage: MutableState<Boolean>,
     selectedTab: MutableState<Int>,
     selectedItem: MutableState<ItemCardDetail?>,
     showEditDialog: MutableState<Boolean>,
@@ -198,140 +191,110 @@ fun MainContent(
         itemViewModel.addCategoryAttribute(attribute)
     }
 
-    AppScaffold(
-        cardDetails = cardDetails,
-        categories = categories,
-        categoryAttributes = categoryAttributes,
-        onSaveEdit = { itemCardDetail ->
-            itemViewModel.updateItemCardDetail(itemCardDetail)
-            itemViewModel.refreshItems()
-        },
-        onManualAdd = { showAddDialog.value = true },
-        onScanAdd = { /* TODO: */ },
-        onDeleteCard = { itemCardDetail -> itemViewModel.deleteItem(itemCardDetail) },
-        onAddCategory = { categoryName -> itemViewModel.addNewCategory(categoryName) },
-        onDeleteCategory = { categoryId -> itemViewModel.deleteCategory(categoryId) },
-        onLoadCategoryAttributes = { categoryId ->
-            itemViewModel.loadCategoryAttributes(categoryId)
-        },
-        onDeleteAttribute = { attributeId ->
-            itemViewModel.deleteCategoryAttribute(attributeId)
-        },
-        onAddAttribute = onAddAttribute,
-        onAddImage = { itemId, onImageAdded ->
-            onPickImage(itemId, onImageAdded)
-        },
-        onDeleteImage = { itemId, imageId, isCoverImage ->
-            itemViewModel.deleteImage(itemId, imageId, isCoverImage)
-            itemViewModel.refreshItems()
-        },
-        onSetCoverImage = { itemId, imageId ->
-            itemViewModel.updateItemCoverImage(itemId, imageId)
-        },
-        selectedTab = selectedTab,
-        onItemSelect = { item -> selectedItem.value = item }
-    )
-
-    // View Dialog
-    selectedItem.value?.let { item ->
-        if (!showEditDialog.value) {
-            ItemViewDialog(
-                item = item,
-                categoryAttributes = categoryAttributes,
-                onDismiss = { selectedItem.value = null },
-                onEdit = { 
-                    isEditFromView.value = true
-                    showEditDialog.value = true 
-                }
-            )
-        }
-    }
-
-    // Edit Dialog
-    if (showEditDialog.value && selectedItem.value != null) {
-        ItemEditDialog(
-            item = selectedItem.value!!,
+    if (showCreatePage.value) {
+        ItemCreatePage(
             categories = categories,
             categoryAttributes = categoryAttributes,
-            onDismiss = {
-                showEditDialog.value = false
-                if (!isEditFromView.value) {
-                    selectedItem.value = null
-                }
-                isEditFromView.value = false
-            },
-            onSave = { editedItem ->
-                itemViewModel.updateItemCardDetail(editedItem)
-                itemViewModel.refreshItems()
-                showEditDialog.value = false
-                if (!isEditFromView.value) {
-                    selectedItem.value = null
-                }
-                isEditFromView.value = false
-            },
-            onAddImage = {
-                onPickImage(selectedItem.value!!.id) { updatedItem ->
-                    selectedItem.value = updatedItem
-                }
-            },
-            onDeleteImage = { imageId -> 
-                val isCoverImage = selectedItem.value!!.coverImageId == imageId
-                itemViewModel.deleteImage(selectedItem.value!!.id, imageId, isCoverImage)
-                itemViewModel.refreshItems()
-            },
-            onCategorySelected = { categoryId ->
-                itemViewModel.loadCategoryAttributes(categoryId)
-            },
-            onSetCoverImage = { _, imageId -> 
-                itemViewModel.updateItemCoverImage(selectedItem.value!!.id, imageId)
-            }
-        )
-    }
-
-    // Add Dialog
-    if (showAddDialog.value) {
-        itemViewModel.ensureCategoriesLoaded()
-        val emptyItem = ItemCardDetail(
-            id = 0,
-            name = "",
-            categoryId = 0,
-            codeType = null,
-            codeContent = null,
-            codeImageId = null,
-            coverImageId = null,
-            createdAt = System.currentTimeMillis(),
-            updatedAt = System.currentTimeMillis(),
-            category = null,
-            codeImage = null,
-            coverImage = null,
-            images = emptyList(),
-            attributes = emptyList()
-        )
-        ItemEditDialog(
-            item = emptyItem,
-            categories = categories,
-            categoryAttributes = categoryAttributes,
-            onDismiss = { showAddDialog.value = false },
+            onNavigateBack = { showCreatePage.value = false },
             onSave = { newItem ->
                 itemViewModel.addNewItem(newItem)
                 itemViewModel.refreshItems()
-                showAddDialog.value = false
-            },
-            onAddImage = {
-                onPickImage(0) { updatedItem ->
-                    // No need to update state since this is a new item
-                }
-            },
-            onDeleteImage = { imageId ->
-                // No need to handle delete for new item
+                showCreatePage.value = false
             },
             onCategorySelected = { categoryId ->
                 itemViewModel.loadCategoryAttributes(categoryId)
+            }
+        )
+    } else {
+        AppScaffold(
+            cardDetails = cardDetails,
+            categories = categories,
+            categoryAttributes = categoryAttributes,
+            onSaveEdit = { itemCardDetail ->
+                itemViewModel.updateItemCardDetail(itemCardDetail)
+                itemViewModel.refreshItems()
+            },
+            onManualAdd = { showCreatePage.value = true },
+            onScanAdd = { /* TODO: */ },
+            onDeleteCard = { itemCardDetail -> itemViewModel.deleteItem(itemCardDetail) },
+            onAddCategory = { categoryName -> itemViewModel.addNewCategory(categoryName) },
+            onDeleteCategory = { categoryId -> itemViewModel.deleteCategory(categoryId) },
+            onLoadCategoryAttributes = { categoryId ->
+                itemViewModel.loadCategoryAttributes(categoryId)
+            },
+            onDeleteAttribute = { attributeId ->
+                itemViewModel.deleteCategoryAttribute(attributeId)
+            },
+            onAddAttribute = onAddAttribute,
+            onAddImage = { itemId, onImageAdded ->
+                onPickImage(itemId, onImageAdded)
+            },
+            onDeleteImage = { itemId, imageId, isCoverImage ->
+                itemViewModel.deleteImage(itemId, imageId, isCoverImage)
+                itemViewModel.refreshItems()
             },
             onSetCoverImage = { itemId, imageId ->
                 itemViewModel.updateItemCoverImage(itemId, imageId)
-            }
+            },
+            selectedTab = selectedTab,
+            onItemSelect = { item -> selectedItem.value = item }
         )
+
+        // View Dialog
+        selectedItem.value?.let { item ->
+            if (!showEditDialog.value) {
+                ItemViewDialog(
+                    item = item,
+                    categoryAttributes = categoryAttributes,
+                    onDismiss = { selectedItem.value = null },
+                    onEdit = { 
+                        isEditFromView.value = true
+                        showEditDialog.value = true 
+                    }
+                )
+            }
+        }
+
+        // Edit Dialog
+        if (showEditDialog.value && selectedItem.value != null) {
+            ItemEditDialog(
+                item = selectedItem.value!!,
+                categories = categories,
+                categoryAttributes = categoryAttributes,
+                onDismiss = {
+                    showEditDialog.value = false
+                    if (!isEditFromView.value) {
+                        selectedItem.value = null
+                    }
+                    isEditFromView.value = false
+                },
+                onSave = { editedItem ->
+                    itemViewModel.updateItemCardDetail(editedItem)
+                    itemViewModel.refreshItems()
+                    showEditDialog.value = false
+                    if (!isEditFromView.value) {
+                        selectedItem.value = null
+                    }
+                    isEditFromView.value = false
+                },
+                onAddImage = {
+                    onPickImage(selectedItem.value!!.id) { updatedItem ->
+                        selectedItem.value = updatedItem
+                    }
+                },
+                onDeleteImage = { imageId -> 
+                    val isCoverImage = selectedItem.value!!.coverImageId == imageId
+                    itemViewModel.deleteImage(selectedItem.value!!.id, imageId, isCoverImage)
+                    itemViewModel.refreshItems()
+                },
+                onCategorySelected = { categoryId ->
+                    itemViewModel.loadCategoryAttributes(categoryId)
+                },
+                onSetCoverImage = { _, imageId -> 
+                    itemViewModel.updateItemCoverImage(selectedItem.value!!.id, imageId)
+                }
+            )
+        }
     }
 }
